📦 Changed project structure a bit
