--- src/persisters/base/mod.rs
@@ -1,7 +1,0 @@
-//! Base persisters that handle operations.
-
-mod file;
-mod orm;
-
-pub use file::File;
-pub use orm::Orm;
