--- src/persisters/error.rs
@@ -1,46 +1,0 @@
-//! Errors related to the File management.
-
-use std::fmt;
-
-/// Errors related to file and path management.
-#[derive(Debug)]
-pub enum FileError {
-    /// Used for file format related issues.
-    UnsupportedFormat,
-    /// Used when two paths are the same.
-    SamePaths,
-    /// Used for operations where you need an old path to use a new path.
-    NoOldPath,
-    /// Used when a path already exists.
-    PathExists,
-}
-
-impl fmt::Display for FileError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::UnsupportedFormat => write!(f, "Unsupported file format; defaulting to CSV"),
-            Self::SamePaths => write!(f, "Both paths are the same"),
-            Self::NoOldPath => write!(f, "Old path doesn't exists"),
-            Self::PathExists => write!(f, "New path already exists"),
-        }
-    }
-}
-
-
-/// Errors related to databases and connection strings.
-#[derive(Debug)]
-pub enum DbError {
-    /// Used when the provided connection string is not supported.
-    UnsupportedDatabase,
-    /// Used when the provided connection string is incorrect.
-    IncorrectConnectionString,
-}
-
-impl fmt::Display for DbError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::UnsupportedDatabase => write!(f, "Unsupported database; defaulting to Sqlite"),
-            Self::IncorrectConnectionString => write!(f, "The provided connection string is incorrect"),
-        }
-    }
-}
