+++ src/persisters/mod.rs
@@ -1,10 +1,8 @@
 //! This is where all the file related management happens.
 
-mod base;
-pub mod error;
 pub mod fs;
 pub mod db;
 pub mod traits;
 
-pub use base::File;
-pub use base::Orm;
+pub use fs::File;
+pub use db::Orm;
